Azure Consistent Storage change 1. remove command Export-ACSLog 2. Update Quota payload
