--- src/ResourceManager/AzureStackStorage/Commands.AzureStackStorage/Farm/AddFarm_new.cs
@@ -1,87 +1,0 @@
-// ----------------------------------------------------------------------------------
-//
-// Copyright Microsoft Corporation
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// ----------------------------------------------------------------------------------
-
-// ----------------------------------------------------------------------------------
-//
-// Copyright Microsoft Corporation
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// ----------------------------------------------------------------------------------
-
-using System.Globalization;
-using System.Management.Automation;
-using Microsoft.AzureStack.Management.StorageAdmin;
-using Microsoft.AzureStack.Management.StorageAdmin.Models;
-
-namespace Microsoft.AzureStack.Commands.StorageAdmin
-{
-    /// <summary>
-    ///     SYNTAX
-    ///         Add-Farm [-SubscriptionId] {string} [-Token] {string} [-AdminUri] {Uri} [-ResourceGroupName] {string} 
-    ///             [-SkipCertificateValidation] [-FarmName] {string} [-SettingAccessString] {string} [-Region] {string} [ {CommonParameters}] 
-    /// 
-    /// </summary>
-    [Cmdlet(VerbsCommon.Add, Nouns.AdminFarm, SupportsShouldProcess = true)]
-    public sealed class AddAdminFarm : AdminCmdlet
-    {
-        const string ShouldProcessTargetFormat = "farm {0} ";
-
-        /// <summary>
-        ///     Farm identifier
-        /// </summary>
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, Position = 4)]
-        [ValidateNotNull]
-        public string FarmName { get; set; }
-
-        /// <summary>
-        ///     setting access string. ex: file:\\localhost\db1\settings
-        /// </summary>
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, Position = 5)]
-        [ValidateNotNull]
-        public string SettingAccessString { get; set; }
-
-        /// <summary>
-        ///     Location of the farm
-        /// </summary>
-        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, Position = 6)]
-        [ValidateNotNull]
-        public string Location{ get; set; }
-
-        protected override void Execute()
-        { 
-            if (ShouldProcess(string.Format(CultureInfo.InvariantCulture, ShouldProcessTargetFormat, FarmName))){
-                FarmCreateParameters request = new FarmCreateParameters
-                {
-                    Location = Location,
-
-                    Properties = new FarmCreate
-                    {
-                        SettingAccessString = SettingAccessString
-                    }
-                };
-
-                FarmGetResponse response = Client.Farms.Create(ResourceGroupName, FarmName, request);
-
-                WriteObject(response);
-            }
-        }
-    }
-}
