+++ src/ResourceManager/AzureStackStorage/Commands.AzureStackStorage/Model/ResponseBase.cs
@@ -18,49 +18,21 @@ using Microsoft.AzureStack.AzureConsistentStorage.Models;
 
 namespace Microsoft.AzureStack.AzureConsistentStorage.Commands
 {
-    internal class ResponseBase : ResourceBase
+    internal class ResponseBase : BasicResponseBase
     {
         public ResponseBase(ResourceBase resource) : this(resource, "Properties")
         {
-            
+
         }
 
         public ResponseBase(ResourceBase resource, string propertyNameToExpand)
+            : base(resource, propertyNameToExpand)
         {
-            Id = resource.Id;
-            Location = resource.Location;
-            Tags = resource.Tags;
-            Type = resource.Type;
-            Name = resource.Name;
-
             // Id is in the following format:
             // /subscriptions/{subid}/resourceGroups/{resourceGroupName}/providers/{providerName}/farms/{farmId}/nodes/{nodeId}
             string[] ids = resource.Id.Split(new[] { '/' }, StringSplitOptions.RemoveEmptyEntries);
             ResourceGroupName = ids[3];
             FarmName = ids[7];
-
-            Type inputType = resource.GetType();
-            Type outPutType = this.GetType();
-
-            PropertyInfo propertiesPropertyInfo = inputType.GetProperty(propertyNameToExpand);
-
-            if (propertiesPropertyInfo != null)
-            {
-                var propertiesValue = propertiesPropertyInfo.GetValue(resource);
-                PropertyInfo[] properties = propertiesValue.GetType().GetProperties(BindingFlags.Public | BindingFlags.Instance);
-                foreach (var property in properties)
-                {
-                    object propertyValue = property.GetValue(propertiesValue);
-                    if (propertyValue != null)
-                    {
-                        PropertyInfo propertyInfoInResponse = outPutType.GetProperty(property.Name, BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
-                        if (propertyInfoInResponse != null)
-                        {
-                            propertyInfoInResponse.SetValue(this, propertyValue);
-                        }
-                    }
-                }
-            }
         }
 
         public string ResourceGroupName { get; set; }
